# AllAboard_swagger.yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: AllAboard API
host: 'localhost'
schemes:
  - http
...
paths:
...
  /seatType/seatType:
    post:
      tags:
        - seatType
      summary: Add a new seat type
      operationId: addSeatType
      consumes:
        - application/json
      produces: 
        - application/json
      parameters:
        - in: body
          name: seatType
          description: Seat type object 
          required: true
          schema:
            $ref: '#/definitions/SeatType'
      responses:
        '200':
          description: Seat type created successfully
  /seatType/seatType/{seatTypeId}:
    put:
      tags:
        - seatType
      summary: Update a seat type
      operationId: updateSeatType
      parameters:
        - name: seatTypeId
          in: path 
          description: ID of the seat type to update 
          required: true
          type: integer
        - name: seatType
          in: body
          description: Updated seat type object 
          required: true
          schema: 
            $ref: '#/definitions/SeatType'
      responses:
        200:
          description: Seat type updated successfully
        400: 
          description: Invalid request data
        404:
          description: Seat type not found
    
    delete: 
      tags:
        - seatType 
      summary: Delete a seat type 
      operationId: deleteSeatType
      parameters:
        - name: seatTypeId
          in: path
          description: ID of the seat type to delete
          required: true
          type: integer
      responses:
        200:
          description: Seat type deleted successfully
        400:
          description: Invalid request data 
        404:
          description: Seat type not found
...
  /peakTime/peakTime:
    post:
      tags:
        - peakTime
      summary: Add a new peak time
      operationId: addPeakTime
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: peakTime 
          description: Peak time object 
          required: true
          schema:
            $ref: '#/definitions/PeakTime'
      responses:
        '200':
          description: Peak time created successfully
  /peakTime/peakTime/{peakTimeId}:
    put:
      tags:
        - peakTime
      summary: Update a peakTime
      operationId: updatePeakTime
      parameters:
        - name: peakTimeId
          in: path
          description: ID of the peak time to update
          required: true
          type: integer 
        - name: peakTime 
          in: body
          description: Updated peak time object 
          required: true 
          schema: 
            $ref: '#/definitions/PeakTime'
      responses:
        200:
          description: Peak time updated successfully 
        400: 
          description: Invalid request data
        404: 
          description: Peak time not found

    delete: 
      tags:
        - peakTime 
      summary: Delete a peak time
      operationId: deletePeakTime
      parameters:
        - name: peakTimeId
          in: path
          description: ID of the peak time to delete
          required: true
          type: integer
      responses:
        200:
          description: Peak time deleted successfully
        400:
          description: Invalid request data 
        404:
          description: Peak time not found  
...
definitions:
...
  SeatType:
    type: object
    properties:
      seat_type_id:
        type: integer
      seat_type_name:
        type: string
...
  PeakTime: 
    type: object 
    properties: 
      peak_time_id:
        type: integer
      start_time: 
        type: string 
        format: time 
      end_time: 
        type: string
        format: time 
      price_increase_percentage: 
        type: number 
...